Maven and code refactoring
